+++ src/components/index/Experience.tsx
@@ -9,10 +9,10 @@ import "react-vertical-timeline-component/style.min.css";
 export default function Experience({ experiences }: ExpProps): ReactElement {
   return (
     <SectionWrapper id="work">
-      <SectionHead Head="I've Worked on/at :  " SubHead="My experience" />
+      <SectionHead Head="I've Worked at :  " SubHead="My experience" />
       <div className="flex flex-col mt-20">
         <VerticalTimeline>
-          {(experiences ?? []).map((xp: any, i) => (
+          {(experiences ?? []).map((xp, i) => (
             <ExperienceCard key={i} {...xp} />
           ))}
         </VerticalTimeline>
