+++ src/components/ui/head/Head.tsx
@@ -5,9 +5,9 @@ export default function Header(): ReactElement {
   return (
     <Head>
       <meta charSet="UTF-8" />
-      <link rel="icon" type="image/svg+xml" href="/clear_logo.svg" />
+      <link rel="icon" type="image/svg+xml" href="/logo-no-background.svg" />
       <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-      <title> YJ | Porfolio </title>
+      <title> | Porfolio </title>
     </Head>
   );
 }
