+++ src/types/ComponentProps.ts
@@ -41,3 +41,12 @@ export interface HeroProps {
   name?: string;
   charge?: string;
 }
+
+export interface ExperienceCardProps {
+  title?: string;
+  company_name?: string;
+  icon?: string;
+  iconBg?: string;
+  date?: string;
+  points: string[];
+}
